refactor(Views v2 providers): move v1 compat filters in separate provider, clean up code a bit
As suggested by @juanfra in
https://github.com/moderntribe/the-events-calendar/pull/2488#issuecomment-489671098 it would make it
easier to maintain and extend to have v1 compat filters handling living in a separate class.
