+++ src/Tribe/Views/V2/Hooks.php
@@ -180,62 +180,4 @@ class Hooks  extends \tad_DI52_ServiceProvider {
 
 		return $query_vars;
 	}
-
-	/**
-	 * Disconnects some Views v1 filters we'll not need for Views v2 to work.
-	 *
-	 * @since TBD
-	 */
-	protected function handle_v1_filters() {
-		/*
-		 * Depending on the context of the request, this might fire before or after Common did bootstrap.
-		 * Let's handle both cases checking whether Common has already loaded or not.
-		 */
-		if ( ! did_action( 'tribe_common_loaded' ) ) {
-			// Common did not bootstrap yet.
-			add_action( 'tribe_common_loaded', [ $this, 'remove_v1_filters' ] );
-
-			return;
-		}
-
-		// Common did already bootstrap: let's remove the filters now.
-		$this->remove_v1_filters();
-	}
-
-	/**
-	 * Removes a list of Views v1 filters to ensure a "clean slate" to handle requests using Views v2 logic.
-	 *
-	 * This method is meant to fire after Common and `Tribe__Events__Main` did bootstrap.
-	 *
-	 * @since TBD
-	 */
-	public function remove_v1_filters() {
-		$filters_to_remove = [
-			'query_vars'    => [
-				[ 'callback' => [ TEC::instance(), 'eventQueryVars' ] ],
-			],
-			'parse_query'   => [
-				[ 'callback' => [ TEC::instance(), 'setDisplay' ], 'priority' => 51 ],
-				[ 'callback' => [ \Tribe__Events__Backcompat::instance(), 'change_qv_to_list' ], 'priority' => 45 ],
-				[ 'callback' => [ \Tribe__Events__Query::class, 'parse_query' ], 'priority' => 50 ],
-			],
-			'pre_get_posts' => [
-				[ 'callback' => [ \Tribe__Events__Query::class, 'pre_get_posts' ], 'priority' => 50 ],
-			],
-			'posts_results' => [
-				[ 'callback' => [ \Tribe__Events__Query::class, 'posts_results' ], 'priority' => 10 ],
-			],
-			'wp'            => [
-				[ 'callback' => [ \Tribe__Events__Main::instance(), 'issue_noindex' ], 'priority' => 10 ],
-			],
-		];
-
-		foreach ( $filters_to_remove as $tag => $filters ) {
-			foreach ( $filters as $filter_data ) {
-				$callback = $filter_data['callback'];
-				$priority = isset( $filter_data['priority'] ) ? $filter_data['priority'] : 10;
-				remove_filter( $tag, $callback, $priority );
-			}
-		}
-	}
 }
